Moved mustache specific parsing to new file. Begun adding comments
